agents/ingestor/requirements.txt
psycopg2-binary
python-dotenv
python-multipart
PyPDF2
sentence-transformers
